refactor: 세션 체크 훅 적용 및 toast 리다이렉트 처리
